ft_printf over, norm ok, printf_destructor validated ... probably ready to validate. :)

+++ srcs/convert_hexa.c
@@ -6,7 +6,7 @@
 /*   By: rotrojan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/08 18:52:41 by rotrojan          #+#    #+#             */
-/*   Updated: 2020/01/09 12:52:36 by rotrojan         ###   ########.fr       */
+/*   Updated: 2020/01/10 17:12:33 by rotrojan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,9 +63,9 @@ static void		right_padding(t_printf *pf, t_spec *spec, uintmax_t d, char c)
 
 void			convert_hexa(t_printf *pf, t_spec *spec, va_list args)
 {
-	uintmax_t	d;
+	unsigned int	d;
 
-	d = va_arg(args, uintmax_t);
+	d = va_arg(args, unsigned int);
 	if (spec->padding == LEFT_PADDING)
 		left_padding(pf, spec, d);
 	else if (spec->padding == ZERO_PADDING)
